refactor: switch everything to using internal packages
