+++ internal/ci/vulnerability_result_diff.go
@@ -1,8 +1,8 @@
 package ci
 
 import (
+	"github.com/google/osv-scanner/internal/grouper"
 	"github.com/google/osv-scanner/internal/output"
-	"github.com/google/osv-scanner/pkg/grouper"
 	"github.com/google/osv-scanner/pkg/models"
 )
