+++ internal/grouper/grouper_test.go
@@ -3,9 +3,8 @@ package grouper_test
 import (
 	"testing"
 
-	"github.com/google/osv-scanner/pkg/grouper"
-
 	"github.com/google/go-cmp/cmp"
+	"github.com/google/osv-scanner/internal/grouper"
 	"github.com/google/osv-scanner/pkg/models"
 )
