+++ pkg/osvscanner/vulnerability_result_internal_test.go
@@ -3,8 +3,8 @@ package osvscanner
 import (
 	"testing"
 
+	"github.com/google/osv-scanner/internal/config"
 	"github.com/google/osv-scanner/internal/testutility"
-	"github.com/google/osv-scanner/pkg/config"
 	"github.com/google/osv-scanner/pkg/lockfile"
 	"github.com/google/osv-scanner/pkg/models"
 	"github.com/google/osv-scanner/pkg/osv"
